requirement.txt
fastapi
uvicorn[standard]
pymongo
pydantic
python-multipart
numpy
pandas
scikit-learn
streamlit
plotly
requests
pillow
opencv-python
tensorflow
joblib